change some names & make AMP all 0.445
